Matching Modes
Add a new matching mode called "path", to go with the existing (default mode) called "domain". Allows hosting at example.com/a and example.com/b instead of a.example.com and b.example.com if desired. Also allow configs to be in a top-level folder instead of within public (view to force the new location in the future).